【安全】【平台通用】修改合入490170083,ANGLE BlitGL copySubTextureCPUReadback Controlled Heap Overflow via Unsynchronized PACK_ROW_LENGTH

Created-by: zhuzhenlongg
Commit-by: lijianguo;gongyadong
Merged-by: ohci1
Description: ### 相关的Issue


### 原因（目的、解决的问题等）


### 描述（做了什么，变更了什么）
同步chromium ohos 0316-0427 代码

### 测试用例（新增、改动、可能影响的功能）


### 是否需要同步至release（如： 132_trunk_6.0-Release  ... ）分支？


### 是否需要刷新社区文档？


- [ ] 是，刷新社区文档pr链接：
- [x] 否


仅文档更新

See merge request: openharmony-tpc/chromium_third_party_angle!103

+++ src/libANGLE/renderer/vulkan/TextureVk.cpp
@@ -3052,8 +3052,9 @@ angle::Result TextureVk::reinitImageAsRenderable(ContextVk *contextVk, const vk:
         // invalidate must be called after wait for finish.
         ANGLE_TRY(srcBuffer->invalidate(renderer));
 
-        size_t dstBufferSize = sourceBox.width * sourceBox.height * sourceBox.depth *
-                               dstFormat.pixelBytes * layerCount;
+        size_t dstBufferSize =
+            static_cast<size_t>(sourceBox.width) * static_cast<size_t>(sourceBox.height) *
+            static_cast<size_t>(sourceBox.depth) * dstFormat.pixelBytes * layerCount;
 
         // Allocate memory in the destination texture for the copy/conversion.
         uint8_t *dstData = nullptr;
